+++ Source/ASCollectionView.mm
@@ -1585,7 +1585,10 @@ static NSString * const kReuseIdentifier = @"_ASCollectionReuseIdentifier";
 
 - (void)setLeadingScreensForBatching:(CGFloat)leadingScreensForBatching
 {
-  _leadingScreensForBatching = leadingScreensForBatching;
+  if (_leadingScreensForBatching != leadingScreensForBatching) {
+    _leadingScreensForBatching = leadingScreensForBatching;
+    [self _checkForBatchFetching];
+  }
 }
 
 - (CGFloat)leadingScreensForBatching